Xong chức năng tạo nhóm, thêm thành viên, tạo hóa đơn, tính toán nợ

+++ Project_BeFair/src/main/java/khanh/ntu/BF/Repository/ExpenseRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import khanh.ntu.BF.models.Expense;
 
-public interface ExpenseRepository extends JpaRepository<Expense, Integer>{
+public interface ExpenseRepository extends JpaRepository<Expense, Long>{
 	List<Expense> findByGroupId(long groupId);
 }
