+++ Project_BeFair/src/main/java/khanh/ntu/BF/Repository/TravelGroupRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import khanh.ntu.BF.models.TravelGroup;
 
-public interface TravelGroupRepository extends JpaRepository<TravelGroup, Integer>{
+public interface TravelGroupRepository extends JpaRepository<TravelGroup, Long>{
 
 }
